% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listAllCandidates.R
\name{listAllCandidates}
\alias{listAllCandidates}
\title{List all available candidates}
\usage{
listAllCandidates(x = NULL, ...)
}
\arguments{
\item{x}{NULL by default, but can be a (path to a) file, in which case the result is written to
that file.}

\item{...}{Arguments passed to \link[base:writeLines]{writeLines}.}
}
\description{
Function will return names of candidate.s
}
\author{
Roman Lustrik {roman.lustrik@biolitika.si}
}

