% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStatement.R
\name{writeStatement}
\alias{writeStatement}
\title{Write statement of candidate to a file.}
\usage{
writeStatement(x)
}
\arguments{
\item{x}{Character. Name of the candidate in form "SURNAME NAME". To see the list of candidates, see
\link[zakonodaja:listAllCandidates]{listAllCandidates}}
}
\value{
Function side-effects are a file written to disk, returns statements as a list.
}
\description{
For a given candidate, the function will read all the statements and output them to a file, one statement per line.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}

